Title: 합과 평균, Time: 6ms, Memory: 0MB, Status: Passed - CodeTree

+++ 250227/합과 평균/sum-and-avg.cpp
@@ -6,6 +6,6 @@ int main() {
     cin >> A >> B;
     cout << fixed;
     cout.precision(1);
-    cout << A+B << " " << (A+B)/2;
+    cout << A+B << " " << (double)(A+B)/2;
     return 0;
 }
